+++ index/secondary_index_manager.hh
@@ -108,6 +108,12 @@ public:
             const std::vector<::shared_ptr<cql3::statements::index_target>> &targets, const gms::feature_service& fs,
         const data_dictionary::database& db) const = 0;
     virtual utils::UUID index_version(const schema& schema) = 0;
+
+protected:
+    std::optional<cql3::description> describe_with_target(
+            const index_metadata& im,
+            const schema& base_schema,
+            const sstring& target_cql) const;
 };
 
 struct stats {
